Initial commit of Accessible Stardew Setup
This commit introduces the initial codebase for the Accessible Stardew Setup, a wxPython-based graphical installer designed to streamline the installation of Stardew Valley accessibility mods, including Stardew Access and it's dependencies (SMAPI, Kokoro, and ProjectFluent).

The project aims to simplify the setup process for players requiring screen reader accessibility enhancements for Stardew Valley, ensuring that the game is more accessible to the blind and low vision communities. This setup tool automatically handles the downloading and installation of necessary mods and dependencies directly from their respective GitHub repositories.

Included in this initial setup:
- Screen reader accessible wxPython GUI framework.
- Automatic handling of mod download and installation
- Ability to select mod variants, such as developer mode smapi, debug Stardew Access, or prerelease builds.
- Optional automatic addition of steam Launch Options for windows steam users
- Optional addition of convenience shortcuts to the user's desktop, including shortcuts to the Stardew Valley Mods folder, official Wiki page, and installed Stardew Access' documentation
- Attempted basic support for macOS and linux users; though this needs testing.

Many thanks to @pepperTheVixen for help with the documentation, and to @pepperTheVixen, @lucieG0, and our discord community for initial suggestions and ideas.